refactor: view 에 필요한 스펙이므로 dto 패키지 이동 - 정석은 각 레이어마다 dto 생성하는 것이 맞다.
