+++ src/main/java/racingcar/step4/docs/README.md
@@ -66,3 +66,13 @@
   - 팀원들이 코드 보는 시간을 줄여줄 수 있고, 비즈니스 흐름 확인도 할 수 있으니까~~
 - [X] if 사용 시, 예외 상황 체크 시에 사용하자
   - ex) RandomCarMoveStrategy
+- [X] input -> controller <-> view 로 데이터 전송 시, dto 위치 질문 드렸는데 말씀주셨던 내용 
+  - 정석을 말씀드리자면, controller, view 모두 dto 를 만들어야합니다. 
+  - 서로 다른 레이어에 데이터를 전송하기 위함이니 <br /> 
+    **각 레이어에 맞는 dto 를 정의하고 사용해야하죠!** (머리 탁!)
+
+  - 그런데, 아마도 동일한 스펙을 가진 클래스가 2개로 중복으로 발생할거에요! <br />
+    서로 다른 레이어이기 때문에 사실상 중복은 아니지만 중복이라 보여질 수 있습니다. <br />
+    그래서 트레이드오프 영역으로 보고 하나의 dto 로 가져가곤 하는데요!
+  - 개인적으로 view 스펙에 필요한 dto 이니 <br />
+    view 패키지에 두고 controller 에서 사용하는 개념으로 트레이드오프 할것 같습니다!
